Break out limits dialog, fix mypy issue with .save()
