+++ nurse/common.py
@@ -12,28 +12,15 @@ guicolors = {
     "DISCON": QtGui.QColor(0, 0, 200),
 }
 
-prefill = [
-    "Room 342, Darth Vader, AGE 45",
-    "Room 123, Frodo Baggins, AGE 50",
-    "Room 243, Tom Cat, AGE 79",
-    "Room 324, Jerry Mouse, AGE 82",
-    "Room 432, Mary Poppins, AGE UNKOWN",
-    "Room 654, Willow Ufgood, AGE 56",
-    "Room 102, Hermione Granger, AGE 17",
-    "Room 124, UNKNOWN, AGE 60",
-    "Room 125, Gandalf the Grey, AGE 63",
-    "Room 164, Luke Skywalker, AGE 43",
-    "Room 167, Indiana Jones, AGE 82",
-    "Room 169, Wonder Woman, AGE 34",
-    "Room 180, Willy Wonka, AGE 69",
-    "Room 181, Thor, AGE 700",
-    "Room 182, Beaver Cleaver, AGE 62",
-    "Room 183, Ebeneezer Scrooge, AGE 99",
-    "Room 002, Inigo Montoya, AGE 45",
-    "Room 184, Ru N. Ning, AGE 43",
-    "Room 185, O. U. Tof, AGE 50",
-    "Room 186, Good Names, AGE 77",
-]
+INFO_STRINGS = {
+    "Avg Flow": ".0f",
+    "Avg Pressure": ".0f",
+    "RR": ".0f",  # (breaths/min)
+    "TVe": ".0f",  # (mL)
+    "PIP": ".0f",  # (cm H2O)
+    "PEEP": ".0f",  # (cm H2O)
+    "I:E time ratio": ".1f",
+}
 
 
 class GraphInfo:
